AFL/requirements.txt
selenium==4.5.0
webdriver-manager==3.5.2
requests==2.28.1
python-dotenv==0.21.0
